+++ src/modules/core.ui.js
@@ -78,9 +78,9 @@
     if (text.match(/<svg/)) className += ' icon';
     let paramAttribute = '';
     if (payload) {
-      if (typeof payload === 'object') paramAttribute = ` data-params="---enc:${tw.core.common.encoder(JSON.stringify(payload))}"`;
-      else if (typeof payload === 'string') paramAttribute = ` data-param="---enc:${tw.core.common.encoder(payload)}"`;
-      else return '<span class="error">ERROR: Button Payload is not a string!</span>';
+      if (typeof payload === 'object') payload = JSON.stringify(payload);
+      else if (typeof payload !== 'string') payload = String(payload);
+      paramAttribute = ` data-params="---enc:${tw.core.common.encoder(payload)}"`;
     }
     return `<button${id ? ' id="' + id + '"' : ''} class="${className}" data-msg="${message}" ${paramAttribute} ${attr}>${text}</button>`;
   }